+++ imports.rb
@@ -7,7 +7,6 @@ require_relative 'art/title_screen'
 require_relative 'assets/ascii'
 require_relative 'assets/cheat_mode'
 require_relative 'assets/names'
-require_relative 'assets/white_space'
 require_relative 'combat/attack_mode'
 require_relative 'combat/beatdown'
 require_relative 'combat/boss_band'
@@ -19,6 +18,7 @@ require_relative 'display/menu_games'
 require_relative 'display/menu_main'
 require_relative 'display/stats'
 require_relative 'display/user_interface'
+require_relative 'display/white_space'
 require_relative 'main/big_boss'
 require_relative 'main/escape_room'
 require_relative 'main/war_machine'
